Add survival guide to user guide

+++ app/src/main/java/com/kylecorry/trail_sense/tools/whitenoise/infrastructure/PinkNoiseGenerator.kt
@@ -41,7 +41,7 @@ class PinkNoiseGenerator {
             SolMath.map(it, min, max, -1.0, 1.0)
         }.toMutableList()
 
-        return soundGenerator.getSound(sampleRate, durationSeconds) {
+        return soundGenerator.getSound(sampleRate, durationSeconds.toFloat()) {
             noise[it + sampleRate]
         }
     }
